refactor: export version helpers for testability
- Export filterPropertiesFiles, filterConnectFiles, extractLatestPropertiesVersion,
  extractLatestConnectVersion from set-latest-version.js
- Update tests to import actual implementations instead of re-implementing
- Addresses CodeRabbit review comment about test duplication
