+++ extensions/version-fetcher/set-latest-version.js
@@ -260,3 +260,49 @@ module.exports.register = function ({ config }) {
     });
   }
 };
+
+// Export helper functions for testing
+module.exports.filterPropertiesFiles = function(files) {
+  return files.filter(f => {
+    if (!f?.src?.stem || !f?.src?.component) return false;
+    const isStreamingComponent = f.src.component === 'streaming' || f.src.component === 'ROOT';
+    const isReferenceModule = f.src.module === 'reference';
+    const isPropertiesFile = f.src.stem.startsWith('redpanda-properties-');
+    return isStreamingComponent && isReferenceModule && isPropertiesFile;
+  });
+};
+
+module.exports.filterConnectFiles = function(files) {
+  return files.filter(f => {
+    if (!f?.src?.stem || !f?.src?.component) return false;
+    const isConnectComponent = f.src.component === 'connect' || f.src.component === 'redpanda-connect';
+    const isComponentsModule = f.src.module === 'components';
+    const isConnectFile = f.src.stem.startsWith('connect-');
+    return isConnectComponent && isComponentsModule && isConnectFile;
+  });
+};
+
+module.exports.extractLatestPropertiesVersion = function(files, semver) {
+  const versions = files
+    .map(f => f.src.stem.replace('redpanda-properties-', ''))
+    .filter(v => {
+      const normalized = v.replace(/^v/, '');
+      return semver.valid(normalized) && semver.prerelease(normalized) === null;
+    });
+
+  if (versions.length === 0) return null;
+
+  return versions.sort((a, b) =>
+    semver.rcompare(a.replace(/^v/, ''), b.replace(/^v/, ''))
+  )[0];
+};
+
+module.exports.extractLatestConnectVersion = function(files, semver) {
+  const versions = files
+    .map(f => f.src.stem.replace('connect-', ''))
+    .filter(v => semver.valid(v) && semver.prerelease(v) === null);
+
+  if (versions.length === 0) return null;
+
+  return versions.sort((a, b) => semver.rcompare(a, b))[0];
+};
